Implements getTokens method
